Query: use `from` trait for concrete impl, views & generic
The `from` trait is implemented to allow switching from a `QueryView` to a generic `Query` either with an ownership transfer, by reference or by mutable reference. The same is applicable from a concrete `Query` to a `QueryView`. From a generic `Query` to the matching `QueryView`, only ownership transfer is allowed because the generic query needs to be moved into the target concrete Query.

Functions using queries as an argument should be able to receive a `QueryView` and either deal with concrete types directly. Such a function could alternatively switch to the generic `Query`. One use case would be accessing the `ffi::GstQuery` from the `QueryView` without the need to use a `match` statement and handle all cases. Another use case would be updating the `Query`'s `Structure`, whatever the concrete query.

`into_ptr` is also implemented on the concrete queries in order to be able to communicate directly with the C API, e.g. in `gst-plugin-rs`.

See unit tests for illustrations of the use cases.