Query: fix mutating queries with ref_count > 1
Previous version could allow extracting a `Query` from a `QueryView` and passing it as mutable even if the underlying `Query` was already immutably borrowed. The solution here is to wrap the `QueryView` in a`QueryWrapper`. The `QueryWrapper` controls the mutability of the `QueryView` based on the status of inner `Query`. Once wrapped in the `QueryWrapper`, the underlying `Query` ownership can't be transfered anymore and mutability controls are handled by the borrow checker.

`GstRc<QueryRef>` is now aliased as `QueryRc` (previously `Query`) as it is used internally. Its role as a mutability controller is mainly handled by `QueryWrapper` in this architecture. `QueryRef` is now named `Query` as it appears in most of the methods.

This approach allows seamlessly handling concrete queries as generic queries and it gives back the ownership and mutability controls to the borrow checker. Except for the increased code base, the main disadvantage with previous architecture is that it is not possible to get back to the `QueryView` from a reference to a `Query` since this operation consumes the `Query`.

These are 4 abstractions and how they must be used:

- `QueryWrapper` must be used at the interface between a code in another language (i.e. when the incoming `ffi::GstQuery`'s `ref_count` might not be 1). This entry point prevents the developper from accessing the concrete query as mutable if he is not allowed to do so. For ease of use, `&QueryWrapper` can `deref` to `&QueryView` and `&Query`. Ideally, the `QueryWrapper` should be the only component able to build a `Query` from `ffi::GstQuery` (this is currently handled by `GstRc`).
- `QueryView` must be used when a function needs to deal with the concrete query variantes. If the underlying `ffi::GstQuery` was received from a source where the `ref_count` might not be 1, `QueryView` will be obtained from a `QueryWrapper` by `ref` or via the function `try_view_mut`. If the underlying `ffi::GstQuery` was generated using Rust's `gst::Query`, the `QueryView` ownership and mutability is controlled by the borrow checker. `&QueryView` can `deref` to `&Query` and `&mut QueryView` can `deref_mut` to `&mut Query`.
- concrete queries (gst::query::Position, ...) are obtained from the `QueryView` enum as a variante when received from the outside or created in place using the appropriate `Query::new...` function. In the first case, they can only be used as `ref` if the underlying `ffi::GstQuery`'s 'ref_count' is > 1 and also as `ref mut` depending on current borrow status. In the second case, the ownership and mutability are controlled by the borrow checker.
- `Query` is never instantiated, but can be used to receive a concrete query, a `QueryView` or a `QueryWrapper` and handle generic functions such as those related to the inner `Structure` or getting the query as a pointer to be passed to a C function.