Comment on why we're not using the test harness
It seems easier to just list the test functions manually and not use the
test harness vs. having to do something smart about the initialization
of gstreamer and the integration.

In addition we expect tests to run in a single thread so that calls to
`integrate` and `disintegrate` do not interfere between tests, as would
be the case for e.g. `test_disintegration` test.