Implement handling of event interests
This was pretty fun a headscratcher. What happens here is that any time
a new Callsite is created (conceptually a location in code which
contains a log statement) we must register ourselves with tracing
runtime and obtain subscribers' interest in the event, and filter
accordingly.

This avoids unnecessarily spending timme on invoking all of the
subscriber machineries among other things. As a future optimisation on
our side, we can also check the interest before allocating/leaking the
`Metadata` and `Callsite`, at the cost of checking the interest twice
when we do actually register the callsite.