Reduce the number of allocations for callsites
With slightly increased amount of unsafe code we can reduce the number
of allocations we make for each callsite to approximately one. It also
helped to simplify the state around metadata setting significantly, and
now a fully valid `GstCallsite` can be created wtihin the same function,
reducing scope of the unsafety to a single function, rather than an
entire module.