Address breaking changes in tracing_core
tracing-core 0.1.22 introduced an unfortunate breaking change. One of
the public types was intended to be non-exhaustive, but turns out
`rustc` was smart enough to make it exhaustive anyway.

The hope was that there wouldn't be any way to obtain the `Kind` from a
`Metadata` and even less of a reason to actually match on one. Alas, it
turned out to be not true for us, where we were trying to open code an
implementation of `PartialEq for Kind`.

In retrospective we should've been using our own defined type anyway and
deriving `PartialEq` as that's just cleaner (we don't need to think
about any future variants added to `Kind`.)

Fixes #6